+++ easygcm-sample/src/main/java/eu/inloop/easygcm/sample/App.java
@@ -2,24 +2,25 @@ package eu.inloop.easygcm.sample;
 
 import android.app.Application;
 import android.os.Bundle;
+import android.util.Log;
 
 import eu.inloop.easygcm.GcmListener;
-import eu.inloop.easygcm.WakeLockRelease;
 
 public class App extends Application implements GcmListener {
 
+    private static final String TAG = "App";
+
     @Override
-    public void onMessage(String s, Bundle bundle, WakeLockRelease wakeLockRelease) {
-        System.out.println("### message: " + s);
-        System.out.println("### bundle:");
-        for (String key: bundle.keySet()) {
-            System.out.println("> " + key + ": " + bundle.get(key));
+    public void onMessage(String from, Bundle data) {
+        Log.v(TAG, "### message from: " + from);
+        Log.v(TAG, "### data: " + from);
+        for (String key : data.keySet()) {
+            Log.v(TAG, "> " + key + ": " + data.get(key));
         }
-        wakeLockRelease.release();
     }
 
     @Override
     public void sendRegistrationIdToBackend(String registrationId) {
-        
+        Log.v(TAG, "### registration id: " + registrationId);
     }
 }