+++ CHANGELOG.md
@@ -1,8 +1,60 @@
-## 1.3.0-SNAPSHOT
+## 1.5.3 (2015-10-13)
+
+Bugfixes:
+
+  - [#45 The package name of permission.C2D_MESSAGE still wrong] (https://github.com/inloop/easygcm/issues/45)
+
+## 1.5.2 (2015-09-28)
+
+Features:
+
+  - [#56 Updated to Google Play Services 8.1.0] (https://github.com/inloop/easygcm/issues/56)
+
+## 1.5.1 (2015-09-22)
+
+Features:
+
+  - [#47 Updated to Google Play Services 7.8.0] (https://github.com/inloop/easygcm/issues/47)
+  - [#48 Used GoogleApiAvailibility instead of GooglePlayServicesUtil] (https://github.com/inloop/easygcm/issues/48)
+
+Bugfixes:
+
+  - [#43 Removed GET_ACCOUNTS permission] (https://github.com/inloop/easygcm/issues/43)
+
+## 1.5.0 (2015-07-02)
 
 This release brings incompatible changes. Please README and sample project for details on integration of the library in your project. Shortly:
 
-  - onMessage is changed, you need to call `wakeLockRelease.release()` after your job is done
+  - `onMessage` is changed, no need to release a wake lock. Make sure, your blocking tasks are executed directly in `GcmListener` callback methods, as they run in background thread. Creating a new background task would not be guaranteed to run.
+  - no need to add custom manifest placeholder `localApplicationId`
+
+Features:
+
+  - [#33 Support for Google Play Services 7.5](https://github.com/inloop/easygcm/issues/33)
+  - [#34 Switch lib to use ${applicationId} in manifest once Android bug is fixed](https://github.com/inloop/easygcm/issues/34)
+  - GCM sender ID is now defined using Google Services Gradle plugin via `google-services.json` as described in [Implementing GCM Client on Android ](https://developers.google.com/cloud-messaging/android/client)
+
+Bugfixes:
+
+  - [#32 BroadcastReceiver trying to return result during a non-ordered broadcast](https://github.com/inloop/easygcm/issues/32)
+  - [#36 BroadcastReceiver trying to return result during a non-ordered broadcast](https://github.com/inloop/easygcm/issues/36)
+
+## 1.4.0 (2015-05-14)
+
+Features:
+
+  - [#26  Allow custom Play Services handling and re-registration](https://github.com/inloop/easygcm/pull/26)
+  - Update Google Play Services to 7.3
+  
+Bugfixes:
+
+  - [#30 Fix build scripts to use local project instead of build from repository](https://github.com/inloop/easygcm/pull/30)
+  
+## 1.3.0 (2015-03-12)
+
+This release brings incompatible changes. Please README and sample project for details on integration of the library in your project. Shortly:
+
+  - `onMessage` is changed, you need to call `wakeLockRelease.release()` after your job is done
   - `GcmHelper.init()` is changed, sender ID is now configured in XML instead of passing as parameter
   - you need to add custom manifest placeholder `localApplicationId` in your build script to workaround Android bug
 
@@ -16,6 +68,8 @@ Bugfixes:
 
 Features:
 
+  - [Support for setting custom GcmListener](https://github.com/inloop/easygcm/pull/21)
+  - [Retry if GCM registration fails](https://github.com/inloop/easygcm/pull/20)
   - add option to disable logging, see `GcmHelper.setLoggingEnabled()`
 
 ## 1.2.3 (2014-12-22)